pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "marketmind"
version = "0.1.0"
description = "Multi-agent marketing intelligence system"
requires-python = ">=3.11,<3.13"

dependencies = [
    # LLM & Agents
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.2",
    "langchain-community>=0.3",
    "langgraph>=0.2",

    # RAG & Vector DB
    "pinecone>=5.0",
    "sentence-transformers>=3.0",
    "rank-bm25>=0.2",
    "tiktoken>=0.7",

    # API & UI
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "streamlit>=1.40",

    # Database
    "psycopg2-binary>=2.9",
    "pymongo>=4.8",
    "pgvector>=0.3",
    "sqlalchemy>=2.0",

    # MLOps & Observability
    "mlflow>=2.16",
    "langsmith>=0.1",

    # Utilities
    "python-dotenv>=1.0",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "httpx>=0.27",
    "tenacity>=9.0",
    "mcp[cli]>=1.0",
    "fpdf2>=2.7",
    "pypdf>=4.0",
    "pandas>=2.1",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
    "mypy>=1.11",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
